+++ middlewares/uploadMiddleware.ts
@@ -1,0 +1,13 @@
+import multer from "multer";
+import path from "path";
+
+const storage = multer.diskStorage({
+    destination: (req, file, cb) => {
+        cb(null, 'Images')
+    },
+    filename: (req, file, cb) => {
+        cb(null, Date.now()  + path.extname(file.originalname))
+    }
+});
+
+export const upload = multer({ storage: storage });
